S81 : Add Black bandeau + move HowTo component
